refactor(tests/cli/execution): moving to typescript
